Decent Update
Worked on getting the ammopool system in, worked on the basics of the weapon system and splitting the weapons up into new  scriptable objects to make it easy to create and expand on weapons, did the same thing for bullets. Made the controls for the new character controller, and have it linked accordingly, added in the setup for a base weapon container with the location and muzzle point and cleaned some things up.